CA0/producer/requirements.txt
kafka-python==2.0.2
pandas==2.1.3
numpy==1.25.2
python-json-logger==2.0.7